Move tests to app level as they are not working in a library module
