+++ app-k9mail/src/test/kotlin/app/k9mail/DependencyInjectionTest.kt
@@ -1,4 +1,4 @@
-package com.fsck.k9
+package app.k9mail
 
 import android.view.ContextThemeWrapper
 import androidx.lifecycle.LifecycleOwner
@@ -29,10 +29,8 @@ import org.mockito.kotlin.mock
 import org.openintents.openpgp.OpenPgpApiManager
 import org.robolectric.RobolectricTestRunner
 import org.robolectric.RuntimeEnvironment
-import org.robolectric.annotation.Config
 
 @RunWith(RobolectricTestRunner::class)
-@Config(application = TestApp::class)
 class DependencyInjectionTest : AutoCloseKoinTest() {
     private val lifecycleOwner = mock<LifecycleOwner> {
         on { lifecycle } doReturn mock()
