+++ app-k9mail/src/test/kotlin/app/k9mail/unread/UnreadWidgetDataProviderTest.kt
@@ -1,4 +1,4 @@
-package com.fsck.k9.widget.unread
+package app.k9mail.unread
 
 import android.content.Context
 import assertk.assertThat
@@ -6,7 +6,6 @@ import assertk.assertions.isEqualTo
 import assertk.assertions.isNull
 import com.fsck.k9.Account
 import com.fsck.k9.Preferences
-import com.fsck.k9.TestApp
 import com.fsck.k9.controller.MessageCounts
 import com.fsck.k9.controller.MessageCountsProvider
 import com.fsck.k9.mailstore.Folder
@@ -15,6 +14,8 @@ import com.fsck.k9.mailstore.FolderType
 import com.fsck.k9.search.SearchAccount
 import com.fsck.k9.ui.folders.FolderNameFormatter
 import com.fsck.k9.ui.messagelist.DefaultFolderProvider
+import com.fsck.k9.widget.unread.UnreadWidgetConfiguration
+import com.fsck.k9.widget.unread.UnreadWidgetDataProvider
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.koin.test.AutoCloseKoinTest
@@ -22,10 +23,8 @@ import org.mockito.kotlin.doReturn
 import org.mockito.kotlin.mock
 import org.robolectric.RobolectricTestRunner
 import org.robolectric.RuntimeEnvironment
-import org.robolectric.annotation.Config
 
 @RunWith(RobolectricTestRunner::class)
-@Config(application = TestApp::class)
 class UnreadWidgetDataProviderTest : AutoCloseKoinTest() {
     private val context: Context = RuntimeEnvironment.getApplication()
     private val account = createAccount()
